+++ packages/polaris-viz/src/components/DonutChart/stories/CustomArc.stories.tsx
@@ -16,8 +16,8 @@ const CustomArcTemplate: StoryFn<DonutChartProps> = (args: DonutChartProps) => {
           Default: {
             arc: {
               cornerRadius: 5,
-              thickness: 50,
-            }
+              thickness: 18,
+            },
           },
         }}
       >