+++ packages/polaris-viz/CHANGELOG.md
@@ -5,7 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-<!-- ## Unreleased -->
+## Unreleased
+
+### Changed
+
+-  Changed `<DonutChart />` to not show 0 or negative values.
+-  Changed `<DonutChart />` to only show a maximum of 5 data points.
+-  Updated `<DonutChart />` default styles.
 
 ## [9.10.6] - 2023-08-24